+++ Source/ACE/Command/Handlers/AdminCommands.cs
@@ -1504,21 +1504,21 @@ namespace ACE.Command.Handlers
             // TODO: output
         }
 
-        // AddCoin [value]
-        [CommandHandler("addcoin", AccessLevel.Player, CommandHandlerFlag.RequiresWorld, 1,
-            "Gives player.",
-            "money (int)")]
-        public static void HandleAddCoin(Session session, params string[] parameters)
-        {
-            try
-            {
-                uint money = uint.Parse(parameters[0]);
-                session.Player.AddCoin(money);
-            }
-            catch (Exception)
-            {
-                // Do Nothing
-            }
-        }
+        ////// AddCoin [value]
+        ////[CommandHandler("addcoin", AccessLevel.Player, CommandHandlerFlag.RequiresWorld, 1,
+        ////    "Gives player.",
+        ////    "money (int)")]
+        ////public static void HandleAddCoin(Session session, params string[] parameters)
+        ////{
+        ////    try
+        ////    {
+        ////        uint money = uint.Parse(parameters[0]);
+        ////        session.Player.AddCoin(money);
+        ////    }
+        ////    catch (Exception)
+        ////    {
+        ////        // Do Nothing
+        ////    }
+        ////}
     }
 }