+++ changelog.md
@@ -1,4 +1,13 @@
 # ACEmulator Change Log
+
+### 2017-10-11
+[Ripley]
+* Added Burden tracking to Players and Containers.
+* Added CoinValue tracking via objects of WeenieType.Coin to Players and Containers.
+* Added Value tracking to Containers.
+* Commented out areas related to above changes that were not implemented correctly.
+* TODO: Update Value/Burden and track changes as StackSize adjustments occur.
+
 ### 2017-10-06
 [Immortalbob]
 * Added equipmentset.cs to house the equipment set enums for future use.
@@ -72,7 +81,6 @@
 * Merged updates into WorldBase.
 * Fixed saving books to the Shard database.
 * Characters now equip items on world entry. (This will reveal bugs in clothing layer priority and model change issues)
-
 
 ### 2017-09-25
 [Og II]